+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Main tauri.conf.json Config update [PR](https://github.com/ldaIas/oauth-diag-tool/pull/1)\
 - Main - Add authorization server metadata call to oauth clients [PR](https://github.com/ldaIas/oauth-diag-tool/pull/2)
+- Main OAuth Server Overrides (redirect, access & refresh expiry) [PR](https://github.com/ldaIas/oauth-diag-tool/pull/4)
 
 ## [0.0.1] - 2026-04-13
 ### Added